+++ src/components/Movies/Movies.js
@@ -1,5 +1,5 @@
 import React from "react";
-import MovieCard from "./MovieCard";
+import MovieCard from "../MovieCard";
 import { Link } from "react-router-dom";
 import "./Movies.css";
